Remove some compiler warnings

+++ packages/seacas/libraries/ioss/src/cgns/Iocgns_Utils.C
@@ -176,8 +176,7 @@ namespace {
     int    min_proc = -1;
     for (int i = 0; i < static_cast<int>(work.size()); i++) {
       if (work[i] < min_work &&
-          proc_adam_map.find(std::make_pair(zone->m_adam->m_zone, static_cast<int>(i))) ==
-              proc_adam_map.end()) {
+          proc_adam_map.find(std::make_pair(zone->m_adam->m_zone, i)) == proc_adam_map.end()) {
         min_work = work[i];
         min_proc = i;
         if (min_work == 0) {
